pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "z-image-studio"
version = "0.1.0"
description = "Z-Image Studio: local image toolkit (CLI + Web UI)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "diffusers",
    "fastapi>=0.123.0",
    "peft>=0.18.0",
    "platformdirs>=4.0.0",
    "python-multipart>=0.0.20",
    "sdnq",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
    "uvicorn>=0.38.0",
    "mcp",
]

[project.scripts]
zimg = "zimage.cli:main"
zimg-mcp = "zimage.mcp_server:main"

[tool.uv.sources]
diffusers = { git = "https://github.com/huggingface/diffusers" }
sdnq = { git = "https://github.com/Disty0/sdnq" }

[tool.hatch.build.targets.wheel]
packages = ["src/zimage"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

